Standardize Typescript NPM behavior, use NPM pack for E2E & reduce controller image size (#93)
This PR ensures that E2E tests only use bundled NPM artifacts so that if
something is not included on `npm publish`, it will also not be included
in the E2E tests to catch bad publish changes in CI before merging. This
PR also exposes declarations & sourcemaps for the SDK to allow
consistent tooling experience with the SDK while also exposing complete
source code to users when viewing source definitions. Lastly the
Dockefile was updated to strip excess non-javascript files used in
`node_modules` to reduce the controller image size further.