metrics endpoint in pepr (#172)
added a simple metrics route to pepr, not exposed via servicemonitor a
WIP for now.

for now, I've only added 3 counters that live in the /mutate endpoint:
alert: if you use the wrong `API KEY`
summary: it uses a full summary for OK requests. 
error: a counter if there are issues. 

I've not yet figured out the best way to allow extended metrics into
each of the capabilities, but for now, the simplicity of this should
help us drive that.

https://github.com/defenseunicorns/pepr/issues/24

---------

Co-authored-by: Megamind <882485+jeff-mccoy@users.noreply.github.com>