feat: create a library of helper functions that make working with pepr easier (#415)
## Description

Added module-helpers.ts and module-helpers.spec.ts.

module-helpers.ts has a containers function that accepts a request
parameter and an optional containerTypes parameter. The request is a
PeprValidateRequest<a.Pod> that identifies the pod from which a list of
containers should be returned.

If no containerTypes parameter is included, all containers in the pod
will be returned. If "containers" is passed in for the containerTypes,
then only spec.containers will be returned. If "initContainers" is
passed in for the containerTypes, then only spec.initContainers will be
returned. If ephemeralContainers is passed in for the containerTypes,
then only spec.ephemeralContainers will be returned.

In order to keep the flexibility of pulling the module-helpers into a
separate module in the future, the containers function was put into the
new module-helpers.ts file instead of modifying the PeprValidateRequest
class.

## Related Issue

Fixes #
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/pepr/blob/main/CONTRIBUTING.md#submitting-a-pull-request)
followed