fix: watch does not respect ns testing (#398)
## Description

Watch is not respecting namespace criteria. This PR is to make watch
respect the ignoredNamespaces and the capability namespaces. If those
two criteria are not respected, a warning is given to the user.

Bindings must respect ignoreNamespace and capability namespace, if not
the build will not occur.

The `alwaysIgnore.namespaces` is the first priority, then capability
namespaces

## Related Issue

Fixes #395 
<!-- or -->
Relates to #

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/pepr/blob/main/CONTRIBUTING.md#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Barrett <81570928+btlghrants@users.noreply.github.com>