chore: remove ignore labels (#626)
## Description

Ignoring the admission of objects based on given labels is a security
risk. Insecure objects could be admitted into the cluster passing the
Validating Webhook creating risk. Due to that, we are not implementing
ignore labels in Pepr Core.

However, this can be done by the module author. The person creating the
Pepr Module (Admission Controller) may choose to create some sort of
exception/exemption back door for a given use-case, we recommend that
this backdoor is CRD based and RBAC can be applied.

## Related Issue

Fixes #622 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/pepr/blob/main/CONTRIBUTING.md#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Barrett <81570928+btlghrants@users.noreply.github.com>