fix: switch to node-lts chainguard image (#716)
Chainguard stopped publishing versioned images outside of `latest` last
fall and so dependabot never picked up a newer version. Thanks to
@eddiezane pointing this out, we were using a very stale image. This
switches the upstream image to the equivalent lts version of Node, which
will change version in the next couple of months, but matches our
current NodeJS compatibility for Pepr.