feat: watchcfg is configurable for KFC from Pepr (#766)
## Description

Under specific conditions the Kubernetes Watch will stop sending events
to the client even though the connection seems healthy. This issue is
well known. After researching, we found that tuning the resync interval
is the best way to ensure the connection stays healthy.

Note that this issue is very rare and is not likely to occur but if your
environment has a lot of network pressure and constant churn it is a
possibility.

## Related Issue

Fixes #765 
<!-- or -->
Relates to # https://github.com/kubernetes-client/javascript/issues/596

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/contributor-guide/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>