refactor: run build in container (#826)
## Description

Currently the build process for pepr builds the code outside of the
container and then copies the transpiled code in.

This PR changes the build process to run the build step inside of a
container. This ensures that builds are fully reproducible and not
dependent on external tooling. It also streamlines building pepr
locally.

## Related Issue

## Type of change

- [x] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/contributor-guide/#submitting-a-pull-request)
followed

Signed-off-by: Eddie Zaneski <eddiezane@gmail.com>
Co-authored-by: Case Wylie <cmwylie19@defenseunicorns.com>