feat: add flag for chart based zarf.yaml and update templated fields in chart deployments (#830)
## Description
Updates `npx pepr build` to take in an optional argument `--zarf
chart|manifest` so that the generated zarf yaml will use the generated
chart vs the manifest. Default behavior is still to use the manifest.
[pepr 822](https://github.com/defenseunicorns/pepr/issues/822)

Also adds templating to the pod annotations on the deployments and extra
volume/volumeMounts
in support of [uds-core
417](https://github.com/defenseunicorns/uds-core/issues/417)

I tested the generation locally for both the --zarf flag as well as
rendering the templates with a values file that added pod annotations
and extra volumes. There didn't seems to be any area really covering
testing this right now. The existing tests in helm.test.ts do still
generically check the templates are valid that were updated here.

## Related Issue

Fixes #
[822](https://github.com/defenseunicorns/pepr/issues/822)
Relates to #
[417](https://github.com/defenseunicorns/uds-core/issues/417)

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/contributor-guide/#submitting-a-pull-request)
followed

---------

Co-authored-by: Case Wylie <cmwylie19@defenseunicorns.com>