chore: update error messages (#844)
## Description
We have noticed that where we throw an error and only use e as the
message, we may not display readable and helpful information in the Pepr
logs. Throwing the error with a specific message results in the Pepr
logs displaying helpful information about what went wrong. We should
check Pepr and KFC for additional instances of throwing errors using
just e and update the error messages.
...

## Related Issue

Fixes #843 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [ ] Test, docs, adr added or updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed