chore: clear cache on failure (#866)
## Description

We were seeing behavior where on a `Store.removeItem()`, if the item was
not there the .Patch call returns a 422 but not clear the body of the
patch. This means that every proceeding patch will also fail. This
clears the cache on failure.

## Related Issue

Fixes #865 
<!-- or -->
Relates to #

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>