chore: named callbacks ADR (#676)
## Description

Pepr is a system of callbacks that produce logs. The callbacks are
anonymous, so debugging which log comes from which callback can be
confusing unless you manually write a name in the log message from
within the callback. This consideration needs an ADR.

## Related Issue

Fixes #675 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://github.com/defenseunicorns/pepr/blob/main/CONTRIBUTING.md#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Barrett <81570928+btlghrants@users.noreply.github.com>
Co-authored-by: schaeferka <kim@defenseunicorns.com>