feat: sharded queue (#1025)
## Description

Changes the Queue to a queueRecord<string, Queue> so that we enqueue
items of the same name. This creates multiple queues but cuts down on
reconciliation time. This is based on issue #1019.

e2e test is here -
https://github.com/defenseunicorns/pepr-excellent-examples/pull/52 which
includes the original tests

Think about adding a new metrics for queue, maybe items per queue

## Related Issue

Fixes #1019 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Other (security config, docs update, etc)

## Checklist before merging

- [x] Test, docs, adr added or updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Barrett LaFrance <barrett.t.lafrance@gmail.com>