chore: different periods between send and receive (#1563)
## Description

We lowered the debounce period (to one second) in which our Store
Subscribers **receive** updates from the Watch on the `PeprStore`
resource, meaning they can react to changes faster.

The same debounce period was responsible for **sending** JSON patches to
the `PeprStore` resource. The cache that holds these updates is updated
and the items that were sent are deleted after the JSON patch succeeds.
This meant that if we are sending events every second, and the
kube-apiserver does not process the patch in a second, then the same
events will be sent over multiple intervals. It is unclear if this would
make the data incorrect as it is idempotent but it certainly introduces
pressure on the network which could create a fallout of other network
created problems.

This PR:
- Creates two different debounce periods, one for send and one for
receive
- Sends a reason for the reject in `SetItemAndWait/RemoveItemAndWait`
- Cleans up the timeouts in `SetItemAndWait/RemoveItemAndWait`
- Adds an array item to the JSON patch to the Store to ensure a watch
event occurs to resolve promises in Set/RemoveItemAndWait


e2e: https://github.com/defenseunicorns/pepr-excellent-examples/pull/204
## Related Issue

Fixes #1561
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>