chore: warn devs when their feature branches may be too large (#1571)
## Description

Collectively, the Pepr team has reviewed some large PRs lately (e.g.,
#1262, #1543, #1539, #1396, #1402, #1407). We prefer several smaller PRs
to reduce the amount of time spent in code-review and to encourage
working in small chunks. This script is an example of a pre-push hook
using Husky that warns devs when their branch may be too large.

This PR does not fulfill a need in the backlog, and does not _need_ to
be merged. Rather, it's offered as an example to how a Pepr developer
might use a similar script for their local development.

## Related Issue

None.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed