chore: pull-back & prep for update of TS (after next pepr release) (#1611)
## Description

This PR expands the version of `typescript` that Pepr will accept (by
introducing a pattern to it's version specifier) in order to allow
updating our excellent examples module versions after next release of
Pepr.

## Background

Given how far behind the tip of the `typescript` release stream we are,
a previous PR tried to update it but was unsuccessful because:

- Pepr declares a specific peerDep version of `typescript`
- Pepr-excellent-examples modules declare 1) a prodDep on a pepr
release, and 2) a devDep on a specific version of `typescript`, example:
[here](https://github.com/defenseunicorns/pepr-excellent-examples/blob/main/hello-pepr-alias/package.json).

So, after last update we had a situation where:
- pepr `main` --> typescript 5.7.2
- pepr `v0.42.1` --> typescript 5.3.3
- pexex modules --> typescript 5.7.2

That meant that trying to `npm install` in the excellent examples gave a
dependency conflict, specifically (in this case) between pepr v0.42.1
wanting _only_ typescript@5.3.3 vs the examples wanting (the updated)
typescript@5.7.2.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [ ] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed