chore: remove soak-interrupts.yaml (#1714)
## Description

The intention behind the soak-interrupts test was to test the informer
pattern by disconnecting the watch and using the `LAST_SEEN_TIME` and
relist to trigger trigger fetch calls against watch resources t update
the ephemeral informer cache and dispatch missed events.

There is currently no way to "turn off" the watch connection without
also turning off the "list" which both communicate to the
kube-apiserver. If there were a way it would need to be a new feature in
KFC.

The way this test is written now it scales down the Pepr pod and brings
it back up and expects the events to be processed. This case is tested
in nearly every excellent example that has a watch and in the UDS test.
There is nothing unique about this test.


## Related Issue

Fixes #1708 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>