chore: restore early exit on build no-embed (#1731)
## Description

When running a build using `--no-embed` the build should _not_ generate
the Pepr module as a secret as used during typical builds. As we were
working towards reducing statements in the build file a function was
created to handle the embed option and stop execution. This PR updates
the function to stop execution which was causing the command to write to
`stdErr`

While implementing this fix I encountered errors and needed to satisfy
eslint configs
```bash
src/cli/build.ts
  131:13  warning  Async arrow function has too many statements (21). Maximum allowed is 20  max-statements
  131:13  warning  Async arrow function has a complexity of 11. Maximum allowed is 10        complexity
```


## Related Issue

Fixes #1659 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Barrett <81570928+btlghrants@users.noreply.github.com>
Co-authored-by: Sam Mayer <sam.mayer@defenseunicorns.com>