chore: nightly releases of pepr cli and controller image (#1738)
## Description

As the next step in Pepr's automation journey this PR creates nightly
releases of the CLI and controller image. As discussed, the nightly is
version is as followings, `pepr_version`-night.[0-x]`.

The first time this is run, there will be a nightly of the current
release with a prerelease version of 0 that will increment as many times
as the task is run until the next release comes out, then it will start
over at 0.

The tag is always set to nightly, the latest pepr code can be utilized
through `npx pepr@nightly [command]`


## Related Issue

Fixes #1692 
<!-- or -->
Relates to #1096 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>