chore: add more testing around asset class (#1753)
## Description

The assets class does not have unit tests. The purpose of this PR is to
add coverage to ensure the class is working as expected.

Due to the use of dependency injection we are loosely helper functions
with the method and functions on the assets class that rely on members
of the assets class. Due to that, the intent of this PR is to ensure
those injected functions are being with the correct members of the
assets class and to hedge against accidentally change by ensuring
certain functions are called a certain number of times.

## Related Issue

Fixes #1655 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>
Co-authored-by: Sam Mayer <sam.mayer@defenseunicorns.com>