chore: reduce duplication between pepr-build and pepr-build-wasm in journey tests (#1804)
## Description

The Pepr build journey test had a lot of duplicate functions between
`pepr-build.ts` and `pepr-build-wasm.ts`.

This PR reduces the duplication as much as possible by creating a common
`validateZarfYaml`, `validateClusterRoleYaml`, and
`validateOverridesYamlRbac`. The tests technically test different thing
so without creeping scope this seems to satisfy the requirements.


**Note**:

- During this PR i realized that pepr-build was building twice with
different argumens and operating in different directories, it was
extremely hard to follow. I have cleaned up the test so that it tests
one workflow through the build in the appropriate directories.
- Running the tests locally worked, I was confused, then I realized that
`npm test` did not actually run the wasn't journey tests.

This PR accounts for all of that since it was inscope in order to get
the journeys passing


## Related Issue

Fixes #1344 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>