chore: bump esbuild version (#1807)
## Description

This PR mitigates
[GHSA-67mh-4wv8-2f99](https://vat.dso.mil/vat/finding?name=GHSA-67mh-4wv8-2f99).
We use `esbuild` to bundle the Pepr module, which is during the build
process in the CLI. This code is not accessible during runtime. There is
no attack vector in Pepr vulnerable to this.

Doing this upgrade will avoid us flagging on security scans so that we
do not need to justify a security finding that does not impact us.

This PR also uses parameterized tests to assert on the presence of
expected files after `pepr build --no-embed`.

### Update Consideration

`esbuild:0.25.0` omits legal comment output files when empty, see
[esbuild/#3670](https://github.com/evanw/esbuild/issues/3670).

As a result, pepr now mirrors this behavior with output to `dist/`. It
is unlikely that this will cause issues but is something users should be
aware of, depending upon what they care about in build artifacts.

## Related Issue

Security hotfix.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Co-authored-by: Case Wylie <cmwylie19@defenseunicorns.com>