chore: improve test coverage for webhook generation (#1815)
## Description

We identified a testing gap that exists when a webhook is generated from
an instance of the `Assets` object that has `Assets.host` defined. This
test case addresses that gap.

**NOTE**: There's some tight coupling with the `Assets` class and the
use of a massive JSON string to create our test `assets` object in this
file. Resolving those issue are outside the scope of this PR but do need
to be addressed by future work. See #1816 and #1817

## Related Issue

Fixes #1723

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed