chore: prefer it() over test() for BDD-style test names (#1818)
## Description

Our test suite uses `test()` and `it()` in declaring tests. While
they're aliases for each other and do the same thing, it provides
uncertainty for new contributors on which pattern to use. Additionally,
using `it()` prompts test writers to think in a BDD-style for test
naming.

This PR replaces the use of `test()` with `it()`.

## Related Issue

Hotfix

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

Co-authored-by: Case Wylie <cmwylie19@defenseunicorns.com>