chore(tests): refactor test data generation for RBAC tests (#1898)
## Description

This PR is a continued effort to resolve #1329. It uses a generator
function to create test data.

I had a sync discussion with @cmwylie19 to review how we confidently
moved from the hardcoded data to a generator function. See PR state as
of (4d900364e905763acf292289640dcbcfa6da0375) for the technical details.

## Related Issue

Fixes #1329 
<!-- or -->
Relates to #1878 

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [x] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [x] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed