chore: webhook timer coverage in mutate-processor test (#1897)
## Description

This is very hard to test because the Capability does not have actual
bindings. Because of that the test stops
[here](https://github.com/defenseunicorns/pepr/blob/4e8549a43bb911072bd0316f7dace7172e5eb1a0/src/lib/processors/mutate-processor.ts#L197)
and returns early. The only meaningful tests I could do was to ensure
the timer start and stop were called.

There is already too much mocking in this file I cannot get any
assertions to pass for base64 calls here.

**UPDATE**: was able to validate the first decodeData was called by
shuffling files around and mocking the call in `mutate-process.test.ts`.
The rest of the test is untestable due to no bindings in the
capabilities

Needs https://github.com/defenseunicorns/pepr/pull/1895 to pass.

## Related Issue

Fixes #1892 
<!-- or -->
Relates to #

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Other (security config, docs update, etc)

## Checklist before merging
- [ ] Unit,
[Journey](https://github.com/defenseunicorns/pepr/tree/main/journey),
[E2E Tests](https://github.com/defenseunicorns/pepr-excellent-examples),
[docs](https://github.com/defenseunicorns/pepr/tree/main/docs),
[adr](https://github.com/defenseunicorns/pepr/tree/main/adr) added or
updated as needed
- [ ] [Contributor Guide
Steps](https://docs.pepr.dev/main/contribute/#submitting-a-pull-request)
followed

---------

Signed-off-by: Case Wylie <cmwylie19@defenseunicorns.com>